requirements.txt
streamlit
openai
faiss-cpu
sentence-transformers
scikit-learn
